price cell
